Update gpml-change-upload-directory.php
Good afternoon.
This code works well, but the changed path (/wp-content/uploads/images/id91e8cc636fab491e8cd/file.png ) of the photos is displayed only in the frontend. In the administrative part of the site, in the product card, photos still have paths like /wp-content/uploads/2022/11/file.png

I added an additional code, now in the administrative part in the product card the path to the photos is the same as in the frontend.